+++ packages/query/CHANGELOG.md
@@ -779,6 +779,15 @@
     * @leather.io/stacks bumped to 1.17.2
     * @leather.io/utils bumped to 0.46.3
 
+## [2.51.2](https://github.com/leather-io/mono/compare/@leather.io/query-v2.51.1...@leather.io/query-v2.51.2) (2026-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/stacks bumped to 1.22.0
+
 ## [2.51.1](https://github.com/leather-io/mono/compare/@leather.io/query-v2.51.0...@leather.io/query-v2.51.1) (2026-07-15)