+++ packages/rpc/CHANGELOG.md
@@ -411,6 +411,15 @@
     * @leather.io/stacks bumped to 1.17.3
     * @leather.io/utils bumped to 0.46.4
 
+## [2.24.1](https://github.com/leather-io/mono/compare/@leather.io/rpc-v2.24.0...@leather.io/rpc-v2.24.1) (2026-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/stacks bumped to 1.22.0
+
 ## [2.24.0](https://github.com/leather-io/mono/compare/@leather.io/rpc-v2.23.0...@leather.io/rpc-v2.24.0) (2026-07-15)